requirements.txt
python_dotenv==1.0.1
openai==1.57.3
PyMuPDF==1.25.5

